# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Elemoria Rise of the Pokemon API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /game/player:
    get:
      operationId: getPlayer
      tags: [game]
      summary: Get player data
      responses:
        "200":
          description: Player data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [game]
      summary: Create a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerBody"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /game/player/progress:
    patch:
      operationId: updatePlayerProgress
      tags: [game]
      summary: Update player progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerProgressBody"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /Pokemon/{id}:
    get:
      operationId: getPokemonpecies
      tags: [Pokemon]
      summary: Get a specific Pokemon species
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pokemon species detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pokemonpecies"
  /player/Pokemon:
    get:
      operationId: getPlayerPokemon
      tags: [Pokemon]
      summary: Get all Pokemon owned by the player
      responses:
        "200":
          description: Player's Pokemon
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPokemon"
    post:
      operationId: capturePokemon
      tags: [Pokemon]
      summary: Capture a wild Pokemon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CapturePokemonBody"
      responses:
        "201":
          description: Pokemon captured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPokemon"
  /player/Pokemon/{id}:
    get:
      operationId: getPlayerPokemon
      tags: [Pokemon]
      summary: Get a specific player Pokemon
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player Pokemon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPokemon"
    patch:
      operationId: updatePlayerPokemon
      tags: [Pokemon]
      summary: Update a player Pokemon (rename, evolve, etc)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlayerPokemonBody"
      responses:
        "200":
          description: Updated player Pokemon
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerPokemon"
  /player/party:
    get:
      operationId: getPlayerParty
      tags: [Pokemon]
      summary: Get the player's active battle party (up to 3)
      responses:
        "200":
          description: Player party
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPokemon"
    post:
      operationId: setPlayerParty
      tags: [Pokemon]
      summary: Set the player's battle party
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPartyBody"
      responses:
        "200":
          description: Updated party
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerPokemon"
...
components:
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        avatarColor:
          type: string
        level:
          type: integer
        experience:
          type: integer
        currency:
          type: integer
        currentRegionId:
          type: integer
        totalBattles:
          type: integer
        battlesWon:
          type: integer
        totalPokemonCaptured:
          type: integer
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - avatarColor
        - level
        - experience
        - currency
        - currentRegionId
        - totalBattles
        - battlesWon
        - totalPokemonCaptured
        - badges
        - createdAt
    CreatePlayerBody:
      type: object
      properties:
        name:
          type: string
        avatarColor:
          type: string
        starterPokemonpeciesId:
          type: integer
      required:
        - name
        - avatarColor
        - starterPokemonpeciesId
    UpdatePlayerProgressBody:
      type: object
      properties:
        experience:
          type: integer
        currency:
          type: integer
        currentRegionId:
          type: integer
        totalBattles:
          type: integer
        battlesWon:
          type: integer
        totalPokemonCaptured:
          type: integer
        badges:
          type: array
          items:
            type: string
...
    Pokemonpecies:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        type1:
          type: string
        type2:
          type: string
          nullable: true
        baseHp:
          type: integer
        baseAttack:
          type: integer
        baseDefense:
          type: integer
        baseSpeed:
          type: integer
        baseSpecialAttack:
          type: integer
        baseSpecialDefense:
          type: integer
        evolutionLevel:
          type: integer
          nullable: true
        evolvesIntoId:
          type: integer
          nullable: true
        catchRate:
          type: integer
        spriteUrl:
          type: string
        backSpriteUrl:
          type: string
        moves:
          type: array
          items:
            $ref: "#/components/schemas/Move"
        rarity:
          type: string
      required:
        - id
        - name
        - description
        - type1
        - baseHp
        - baseAttack
        - baseDefense
        - baseSpeed
        - baseSpecialAttack
        - baseSpecialDefense
        - catchRate
        - spriteUrl
        - backSpriteUrl
        - moves
        - rarity
    Move:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        power:
          type: integer
        accuracy:
          type: integer
        pp:
          type: integer
        category:
          type: string
        description:
          type: string
        effect:
          type: string
          nullable: true
      required:
        - id
        - name
        - type
        - power
        - accuracy
        - pp
        - category
        - description
    PlayerPokemon:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        speciesId:
          type: integer
        nickname:
          type: string
          nullable: true
        level:
          type: integer
        experience:
          type: integer
        currentHp:
          type: integer
        maxHp:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        speed:
          type: integer
        specialAttack:
          type: integer
        specialDefense:
          type: integer
        moves:
          type: array
          items:
            $ref: "#/components/schemas/Move"
        isInParty:
          type: boolean
        partySlot:
          type: integer
          nullable: true
        statusCondition:
          type: string
          nullable: true
        species:
          $ref: "#/components/schemas/Pokemonpecies"
        capturedAt:
          type: string
      required:
        - id
        - playerId
        - speciesId
        - level
        - experience
        - currentHp
        - maxHp
        - attack
        - defense
        - speed
        - specialAttack
        - specialDefense
        - moves
        - isInParty
        - species
        - capturedAt
    CapturePokemonBody:
      type: object
      properties:
        speciesId:
          type: integer
        level:
          type: integer
        nickname:
          type: string
      required:
        - speciesId
        - level
    UpdatePlayerPokemonBody:
      type: object
      properties:
        nickname:
          type: string
        currentHp:
          type: integer
        experience:
          type: integer
        level:
          type: integer
        statusCondition:
          type: string
          nullable: true
        evolveToSpeciesId:
          type: integer
    SetPartyBody:
      type: object
      properties:
        PokemonIds:
          type: array
          items:
            type: integer
          maxItems: 6
      required:
        - PokemonIds
...