# lib/api-zod/src/generated/types/aetherSpecies.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Api
 * Elemoria Rise of the Pokemon API
 * OpenAPI spec version: 0.1.0
 */
import type { Move } from "./move";

export interface Pokemonpecies {
  id: number;
  name: string;
  description: string;
  type1: string;
  type2?: string | null;
  baseHp: number;
  baseAttack: number;
  baseDefense: number;
  baseSpeed: number;
  baseSpecialAttack: number;
  baseSpecialDefense: number;
  evolutionLevel?: number | null;
  evolvesIntoId?: number | null;
  catchRate: number;
  spriteUrl: string;
  backSpriteUrl: string;
  moves: Move[];
  rarity: string;
}